refactor(behavior): Move execution nodes to  local execution nofr and remove deprecated classes.
