+++ machine_data_model/behavior/control_flow_node.py
@@ -1,8 +1,6 @@
 from typing import Any
 from abc import ABC, abstractmethod
-from collections.abc import Callable
 from machine_data_model.behavior.control_flow_scope import ControlFlowScope
-from machine_data_model.nodes.data_model_node import DataModelNode
 from machine_data_model.protocols.frost_v1.frost_message import FrostMessage
 
 
@@ -92,85 +90,3 @@ class ControlFlowNode(ABC):
             return False
 
         return self.node == other.node
-
-
-class LocalExecutionNode(ControlFlowNode):
-    """
-    Abstract base class representing a control flow action node in the control flow graph. A control flow action node is a basic unit of the control flow graph that can be executed locally in the context of a control flow scope.
-
-    :ivar node: The identifier of a local node in the machine data model.
-    :ivar _ref_node: The reference to the node in the machine data model.
-    :ivar get_data_model_node: A callable that takes a node identifier and returns the corresponding node in the machine data model.
-    """
-
-    def __init__(self, node: str, successors: list["ControlFlowNode"] | None = None):
-        """
-        Initialize a new CFActionNode instance.
-
-        :param node: The identifier of a local node in the machine data model.
-        :param successors: A list of control flow nodes that are successors of the current node.
-        """
-        super().__init__(node, successors)
-
-        self.node = node
-        self._ref_node: DataModelNode | None = None
-        self.get_data_model_node: Callable[[str], DataModelNode | None] | None = None
-
-    def get_successors(self) -> list["ControlFlowNode"]:
-        """
-        Gets the list of control flow nodes that are successors of the current node.
-
-        :return: The list of control flow nodes that are successors of the current node.
-        """
-        return self._successors
-
-    def is_node_static(self) -> bool:
-        """
-        Check if the node is static. This is used to determine if the reference node
-        can be resolved at creation time or if it needs to be resolved at execution time.
-
-        :return: True if the node is static, otherwise False.
-        """
-        return self.node is not None and not is_variable(self.node)
-
-    def set_ref_node(self, ref_node: DataModelNode) -> None:
-        """
-        Set the reference to the node in the machine data model.
-
-        :param ref_node: The reference to the node in the machine data model.
-        """
-        assert ref_node.name == self.node.split("/")[-1]
-        self._ref_node = ref_node
-
-    def get_ref_node(self) -> DataModelNode | None:
-        """
-        Get the reference to the node in the machine data model.
-
-        :return: The reference to the node in the machine data model.
-        """
-        return self._ref_node
-
-    def _get_ref_node(self, scope: ControlFlowScope) -> DataModelNode | None:
-        """
-        Get the node referenced by the current node. If the node is static, it returns the
-        reference node. Otherwise, it resolves the value of the node in the scope and
-        retrieves the reference node from the machine data model.
-
-        :param scope: The scope of the control flow graph.
-        :return: The node referenced by the current node.
-        """
-
-        if self.is_node_static():
-            return self._ref_node
-        node_path = resolve_value(self.node, scope)
-        assert self.get_data_model_node is not None
-        return self.get_data_model_node(node_path)
-
-    def __eq__(self, other: object) -> bool:
-        if self is other:
-            return True
-
-        if not isinstance(other, LocalExecutionNode):
-            return False
-
-        return super().__eq__(other)
